+++ src/api/user_data.ts
@@ -5,13 +5,16 @@
 import { mastodon } from "masto";
 
 import Account from "./objects/account";
+import MastoApi from "./api";
 import Storage from "../Storage";
 import Toot from "./objects/toot";
 import { AccountNames, StorageKey, TagNames, TootLike, TrendingTag, UserDataSerialized } from "../types";
-import MastoApi from "./api";
 import { sortObjsByProps } from "../helpers/collection_helpers";
 
-const SORT_TAGS_BY = ["numToots" as keyof TrendingTag, "name" as keyof TrendingTag];
+const SORT_TAGS_BY = [
+    "numToots" as keyof TrendingTag,
+    "name" as keyof TrendingTag
+];
 
 interface UserDataCls extends UserDataSerialized {
     isDataStale: () => Promise<boolean>;
@@ -27,7 +30,7 @@ interface UserApiData {
     mutedAccounts: Account[];
     recentToots: Toot[];
     serverSideFilters: mastodon.v2.Filter[];
-}
+};
 
 
 export default class UserData implements UserDataCls {